Add missing host information to the http.Request
The proxy director should inject the host information
to the request object. This information is expected
with remote selenium hubs.

From the http/request go documentation : "For server
request Host specifies the host on which the URL is sought.
Per RFC 2616, this is either the value of the "Host" header
or the host name given in the URL itself"
See : https://golang.org/src/net/http/request.go